+++ docs/changelog.md
@@ -4,7 +4,7 @@ All notable changes to Homeboy CLI are documented in this file.
 
 (This file is embedded into the CLI binary and is also viewable via `homeboy changelog`.)
 
-## Unreleased
+## [0.33.6] - 2026-01-21
 
 ### Added
 - Add `component add-version-target` command for adding version targets without full JSON spec